Separa as funções de Render e Execute do arquivo index/Funções vao para src/Helpers
