+++ composer.json
@@ -1,7 +1,8 @@
 {
     "autoload": {
         "psr-4": {
-            "App\\": "src/"
+            "App\\": "src/",
+            "App\\Helpers\\": "scr/Helpers"
         }
     }
 }
