+++ public/index.php
@@ -3,25 +3,9 @@
 ini_set('display_errors', 1);
 
 include_once "../vendor/autoload.php";
+require "../src/Helpers/Router.php";
 
 $route = explode('?', $_SERVER['REQUEST_URI'])[0];
-
-function render($view)
-{
-    include "../src/views/_partials/header.php";
-
-    //if (usuario logado)
-    include "../src/views/_partials/nav.php";
-    //endif
-
-    include "../src/views/{$view}.php";
-    include "../src/views/_partials/footer.php";
-}
-
-function execute($controller)
-{
-    include "../src/controller/{$controller}.php";
-}
 
 match ($route) {
     '/' => render('home'),
